# Optimise API model discovery speed